Move settings-panel to it's own folder
