+++ src/DetailsView/components/details-view-overlay.tsx
@@ -12,7 +12,7 @@ import { DetailsViewActionMessageCreator } from '../actions/details-view-action-
 import { PreviewFeatureFlagsHandler } from '../handlers/preview-feature-flags-handler';
 import { PreviewFeaturesPanel, PreviewFeaturesPanelProps } from './preview-features-panel';
 import { ScopingPanel, ScopingPanelProps } from './scoping-panel';
-import { SettingsPanel, SettingsPanelDeps, SettingsPanelProps } from './settings-panel';
+import { SettingsPanel, SettingsPanelDeps, SettingsPanelProps } from './settings-panel/settings-panel';
 
 export type DetailsViewOverlayDeps = ContentPanelDeps & SettingsPanelDeps;
