+++ src/tests/unit/tests/DetailsView/components/settings-panel.test.tsx
@@ -1,21 +1,15 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { IChoiceGroupOption } from 'office-ui-fabric-react';
+import { shallow } from 'enzyme';
 import * as React from 'react';
 import { IMock, Mock, Times } from 'typemoq';
-
-import { shallow } from 'enzyme';
 import { FeatureFlags } from '../../../../../common/feature-flags';
 import { UserConfigMessageCreator } from '../../../../../common/message-creators/user-config-message-creator';
 import { BugServicePropertiesMap, UserConfigurationStoreData } from '../../../../../common/types/store-data/user-configuration-store';
 import { DetailsViewActionMessageCreator } from '../../../../../DetailsView/actions/details-view-action-message-creator';
-import { SettingsPanel, SettingsPanelProps } from '../../../../../DetailsView/components/settings-panel';
+import { SettingsPanel, SettingsPanelProps } from '../../../../../DetailsView/components/settings-panel/settings-panel';
 
 type SettingsPanelProtectedClickFunction = (id: string, state: boolean) => void;
-type SettingsPanelProtectedChoiceGroupChangeFunction = (
-    ev?: React.FormEvent<HTMLElement | HTMLInputElement>,
-    option?: IChoiceGroupOption,
-) => void;
 type SettingsPanelProtectedTextFieldChangeFunction = (
     event: React.FormEvent<HTMLInputElement | HTMLTextAreaElement>,
     newValue?: string,
